# PartnerApplicationServices/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PartnerApplicationServices",
    "version": "v1"
  },
  "paths": {
    "/GetFriends": {
      "get": {
        "tags": [
          "Friend"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/AddFriend": {
      "post": {
        "tags": [
          "Friend"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "friendid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/RemoveFriend": {
      "post": {
        "tags": [
          "Friend"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "friendid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/AllUsers": {
      "get": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Register": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegistrationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/CreateAdmin": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/GetUserDetail": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/UpdateUserDetail": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserDetailRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "UserRegistrationRequest": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "expirymode": {
          "type": "string"
        },
        "expirytime": {
          "format": "date-time",
          "type": "string"
        },
        "avatarid": {
          "format": "int32",
          "type": "integer"
        },
        "emailid": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "preferredname": {
          "type": "string"
        },
        "location": {
          "type": "string"
        }
      }
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UpdateUserDetailRequest": {
      "type": "object",
      "properties": {
        "seq": {
          "format": "int32",
          "type": "integer"
        },
        "userid": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "preferredname": {
          "type": "string"
        },
        "avatarid": {
          "format": "int32",
          "type": "integer"
        },
        "location": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}